app/requirements.txt
flask
mysql-connector-python
pymemcache  # This replaces 'memcache'
pika  # For RabbitMQ